Updated config by moving it into totrello. Also updated defn on spec.
